create new function `check_compatibility` (#95)
Create a new function to check if a database is compatible with austraits 3.0 functions

In particular, this function is designed to check a database's metadata to determine if it has been built using `traits.build` versus older versions of AusTraits that were compiled by the internal `austraits.build` workflow. If a version of AusTraits (austraits.build) that is version 4.0 (or older) is detected, messages alert the user to how to download an old version of austraits.

As part of this, the helper function `util_list_to_df2` that converts nested lists to dataframes is added to austraits as an internal function. (Needed to convert specified sections of the databases' metadata into a dataframe).

Includes a test for the new `check_compatibility` function.

Closes issue #93