still a bug...
I've continued to fuss with this function and there are still problems that are baffling me. I did realise that we also needed to collapse, then separate `basis_of_record` since that can vary with value_type.

Also, there are ~50 datasets where info is collapsed: these are the datasets where, for instance, there are mins and maxs for the same observation_id. For instance, `ABRS_1981` which is in the austraits5.0.0-lite, so I switched to using it. But it is huge!

Now, there is one bigger problem to discuss @dfalster. In AusTraits we now have `bin` and `range` as value types and for these, the delimiter is already `--`. So there is nothing to collapse, but the code will seek to separate these. And this will create a problem, since there won't be `--` in the other rows being separated.

I'm not sure why these functions have been passing tests until now, but it is a pretty big rethink to get them running usefully on all datasets.