continued work on unpacking functions
Continued work on unpacking functions - all the individual tibbles with data and metadata now can be cyclically turned into "combined" format and back to "relational".

The test function runs on a diversity of datasets that are missing different pieces of metadata, to ensure empty rows are added as necessary.

Last thing to do - loop over many datasets and recombine them.