most remaining bugs fixed
The tests to do with extract_data are generally working, but there is a weird, oscillating bug with the order of tables in the database list that I can't work out.

And also there is a test on line 198 (expect_equal(trait_subset$traits %>% dplyr::distinct(dataset_id) %>% nrow(), 5)) alternates between giving 5 & 8 as output.

But bugs fixed:
- works with empty tibbles
- works with vectors of values to match