functions now work with identifiers
All functions now work *WITH* an identifiers table, but not all also work *WITHOUT* the identifiers table, so the tests are failing.

And there are now tests *WITH* an identifiers table.

Stopping playing with this until we decide how to work this out - since we need versions of all functions that work with the new traits.build format in order to build the traits.build format (to bind databases).