`extract ` add parameter for exact matches (#156)
* Add option to extract_ functions to only allow exact matches - versus the default which uses partial string matches.

There are instances where partial matches return too much data:

For instance

- searching on trait = "leaf_area" returns data for "leaf_area" and "leaf_area_ratio"
- searching for taxonomic_status = "accepted" also returns data for "genus accepted" and therefore more taxon names than are sought.

The default is still for partial matches, but there is now a parameter "partial_matches_allowed" which can be set to FALSE.

closes issue #150 

* Bug fix in `extract_data` that turned up through AusFizz. 

It was possible for the grouping variables designated in the #Rejoining Contexts section of this function to be identical across multiple context properties (dataset_id, category, link_id, value, description). There were 4 datasets in AusFizz where this was the case. This led to duplicate listings of link_vals and then as a follow-on join_contexts didn't work because there were duplicate link_vals.

closes issue #154