Update load_austraits.R (#171)
two changes:
- for v7 (and future releases) there are multiple data files (flattened
and relational outputs), so load_austraits.R needs to be changed to
ignore the flattened version
- having update = FALSE as the default, means that for people who have
loaded an old version from Zenodo will not get the updated list of
available versions. I think this will create confusion for many people

will add an issue that we need to create a parallel function
`load_austraits_flattened` for loading the flattened data version.